est_weights

An object containing weighted intervention data (est_weights$analysis_data) and a vector of the centered matching variables (est_weights$matching_vars).

est_weights

Format

est_weights$analysis_data is a data frame with 500 rows and 16 variables:

USUBJID

unique patient number

ARM

Name of the arm of the trial the patient was randomized to

AGE

Patient age, years

SEX

Patient gender: 1 for males and 0 for females

SMOKE

Variable to indicate whether the patients smokes: 1 for yes and 0 for no

ECOG0

ECOOG PS: 1 for ECOG PS 0 and 0 for ECOG PS >=1

response

Objective response

Time

Time to overall survival event/censor

Event

Overall survival event = 1, censor = 0

response

Objective response

Age_centered

Patient's age minus average age in the comparator population

Age_squared_centered

Patient's age squared - (Mean age in the target population squared + Standard deviation of age in the target population squared)

Sex_centered

SEX variable minus the proportion of males in the comparator population

Smoke_centered

SMOKE variable minus the proportion of smokers in the comparator population

ECOG0_centered

ECOG0 variable minus the proportion of patients with ECOG0 in the comparator population

wt

Propensity weight assigned to patient

wt_rs

Propensity weight rescaled